Implemented car driver with wheel control.

+++ Unity/DrivingSimulator/Assets/Scripts/DirectionChangedEventArgs.cs
@@ -5,7 +5,7 @@ using System;
 
 public class DirectionChangedEventArgs : EventArgs {
     public float Direction { get; set; }
-    public DirectionChangedEventArgs(int direction)
+    public DirectionChangedEventArgs(float direction)
     {
         Direction = direction;
     }
